refactor: WiseSaying 클래스 별도로 분리
