+++ src/main/java/org/example/dto/WiseSaying.java
@@ -1,0 +1,35 @@
+package org.example.dto;
+
+public class WiseSaying {
+
+    int id;
+    String content;
+    String author;
+
+    public WiseSaying(int id, String content, String author) {
+        this.id = id;
+        this.content = content;
+        this.author = author;
+    }
+
+    public int getId() {
+        return id;
+    }
+
+    public String getContent() {
+        return content;
+    }
+
+    public void setContent(String content) {
+        this.content = content;
+    }
+
+    public String getAuthor() {
+        return author;
+    }
+
+    public void setAuthor(String author) {
+        this.author = author;
+    }
+
+}
